+++ src/components/offer-gallery/offer-gallery.tsx
@@ -1,5 +1,3 @@
-
-
 type OfferGalleryProps = {
     imageGallery: string;
   }
