+++ src/constants.ts
@@ -1,68 +1,7 @@
-import { BookmarkAttributesType, RatingAttributesType } from './types/offer';
-import { TypesPage } from './types/list-offer';
-
+import { TypesPage } from './types/offers';
 
 const previewImage: string = 'https//16.design.htmlacademy.pro/static/hotel/';
 const galleryImages: string = 'img/apartment-';
-
-const HeaderLogoAttributes = {
-  className: 'header__logo',
-  width: 81,
-  height: 41,
-} as const;
-
-const FooterLogoAttributes = {
-  className: 'footer__logo',
-  width: 64,
-  height: 33,
-} as const;
-
-const PlaceCardAttributes = {
-  'cities': {
-    width: 260,
-    height: 200,
-  },
-  'near-places': {
-    width: 260,
-    height: 200,
-  },
-  'favorites': {
-    width: 150,
-    height: 110,
-  },
-} as const;
-
-const BookmarkAttributes:BookmarkAttributesType = {
-  'place-card': {
-    width: 18,
-    height: 19,
-  },
-  'offer': {
-    width: 31,
-    height: 33,
-  },
-} as const;
-
-const RatingAttributes: RatingAttributesType = {
-  'offer': {
-    style: {
-      width: '80%'
-    },
-    ratingVisibility: true
-  },
-  'reviews': {
-    style: {
-      width: '80%'
-    },
-    ratingVisibility: false
-  },
-  'place-card': {
-    style: {
-      width: '100%'
-    },
-    ratingVisibility: false
-  }
-} as const;
 
 const starsTitle = [
   'terrible','badly', 'not bad', 'good', 'perfect'
@@ -170,8 +109,6 @@ const coordinates = [[ 48.85661, 2.351499],	[ 48.868610000000004, 2.342499],	[ 4
 
 export {LOCATIONS, APARTMENT_NAMES, PLACES_OPTIONS,
   CommentLengthLimit, TITLE, previewImage, galleryImages,
-  coordinates, HeaderLogoAttributes,
-  DEFAULT_ACTIVE_LOCATION , DEFAULT_PLACES_OPTIONS, PageNames,
-  SortOptions, PlaceCardAttributes, BookmarkAttributes,
-  RatingAttributes, APARTMENTS_IMAGE_PLACE_CARD, FooterLogoAttributes,
+  coordinates, DEFAULT_ACTIVE_LOCATION , DEFAULT_PLACES_OPTIONS, PageNames,
+  SortOptions, APARTMENTS_IMAGE_PLACE_CARD,
   AppRoute, AuthorizationStatus, starsTitle};
