+++ src/pages/login-page/login-page.tsx
@@ -1,7 +1,7 @@
 import Header from '../../components/header/header';
 import LocationsItemLink from '../../components/locations-item-link/locations-item-link';
 import AuthorizationForm from '../../components/authorization-form/authorization-form';
-import { PageNames } from '../../constans';
+import { PageNames } from '../../constants';
 import { Helmet } from 'react-helmet-async';
