+++ src/pages/main-page/main-page.tsx
@@ -1,39 +1,16 @@
 import Header from '../../components/header/header';
 import LocationsList from '../../components/locations-list/locations-list';
-import SortOffersList from '../../components/sort-offers-list/sort-offers-list';
-import PlaceCardsList from '../../components/places-card-list/places-card-list';
+import SortForm from '../../components/sort-form/sort-form';
+import PlaceCardsList from '../../components/place-card-list/place-card-list';
 import Map from '../../components/map/map';
-import { LOCATIONS, PageNames } from '../../constans';
-import { ListOfferType } from '../../types/list-offer';
-import { Helmet } from 'react-helmet-async';
+import { LOCATIONS, PageNames } from '../../constants';
+import { ListOfferType } from '../../types/offers';
+import MainPageEmpty from '../main-empty-page/main-empty-page';
 
 
 type MainProps = {
 offers: ListOfferType[];
 };
-
-type MainEmptyProps = {
-    location: string;
-  };
-
-function MainPageEmpty({ location }: MainEmptyProps): JSX.Element {
-  return (
-    <div className="cities__places-container cities__places-container--empty container">
-      <Helmet>
-        <title>Главная страница</title>
-      </Helmet>
-      <section className="cities__no-places">
-        <div className="cities__status-wrapper tabs__content">
-          <b className="cities__status">No places to stay available</b>
-          <p className="cities__status-description">
-          We could not find any property available at the moment in {location}
-          </p>
-        </div>
-      </section>
-      <div className="cities__right-section"></div>
-    </div>
-  );
-}
 
 
 function MainPage({offers}:MainProps): JSX.Element {
@@ -54,16 +31,7 @@ function MainPage({offers}:MainProps): JSX.Element {
               <section className="cities__places places">
                 <h2 className="visually-hidden">Places</h2>
                 <b className="places__found">{offersCount} places to stay in Amsterdam</b>
-                <form className="places__sorting" action="#" method="get">
-                  <span className="places__sorting-caption">Sort by</span>&nbsp;
-                  <span className="places__sorting-type" tabIndex={0}>
-                  Popular
-                    <svg className="places__sorting-arrow" width="7" height="4">
-                      <use xlinkHref="#icon-arrow-select"></use>
-                    </svg>
-                  </span>
-                  <SortOffersList />
-                </form>
+                <SortForm />
                 <PlaceCardsList offers = {offers} pageNames= {PageNames.Main}/>
               </section>
               <div className="cities__right-section">
